Introduces ConfirmEmailComponent and adds Suspense-boundary on ConfirmEmailPage
